requirements.txt
accelerate==1.4.0
addict==2.4.0
aiohttp==3.11.14
attrdict==2.0.1
binpacking==1.5.2
charset_normalizer==3.4.1
cpm-kernels==1.0.11
dacite==1.9.2
datasets==3.2.0
einops==0.8.1
fastapi==0.115.8
gradio==3.48.0
importlib_metadata==8.6.1
jieba==0.42.1
matplotlib==3.10.0
modelscope==1.23.2
nltk==3.9.1
numpy==1.26.4
openai==1.67.0
oss2==2.19.1
pandas==2.2.3
peft==0.11.0
pillow==10.4.0
requests==2.32.3
rouge==1.0.1
safetensors==0.5.3
sentencepiece==0.1.96
tensorboard==2.19.0
tiktoken==0.5.2
tqdm==4.67.1
transformers==4.49.0
transformers_stream_generator==0.0.5
trl==0.15
uvicorn==0.34.0
zstandard==0.23.0
torch==2.7.0
torchvision==0.22.0
ms-swift==3.2.0
opencv-python
qwen_vl_utils==0.0.10 # or 0.0.6